server/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-bcrypt
psycopg2-binary